+++ launcher/config.go
@@ -23,36 +23,28 @@ import (
 	"os"
 	"time"
 
+	"github.com/lightstep/otel-launcher-go/pipelines"
 	"github.com/sethvargo/go-envconfig"
-	"go.opentelemetry.io/collector/translator/conventions"
-	hostMetrics "go.opentelemetry.io/contrib/instrumentation/host"
-	runtimeMetrics "go.opentelemetry.io/contrib/instrumentation/runtime"
-	"go.opentelemetry.io/contrib/propagators/b3"
-	"go.opentelemetry.io/contrib/propagators/ot"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric"
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
-	metricglobal "go.opentelemetry.io/otel/metric/global"
-	"go.opentelemetry.io/otel/propagation"
-	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
-	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
-	selector "go.opentelemetry.io/otel/sdk/metric/selector/simple"
 	"go.opentelemetry.io/otel/sdk/resource"
-	"go.opentelemetry.io/otel/sdk/trace"
 	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
-	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/encoding/gzip"
 )
+
+const lightstepAccessTokenHeader = "lightstep-access-token"
 
 type Option func(*Config)
 
 // WithAccessToken configures the lightstep access token
+// remain compatible with the Lightstep-only launcher for now...
 func WithAccessToken(accessToken string) Option {
+	// this will actually get used
 	return func(c *Config) {
-		c.AccessToken = accessToken
+		// duplicating this isn't great but for now let's just get it working
+		if c.Headers == nil {
+			c.Headers = make(map[string]string)
+		}
+		c.Headers["lightstep-access-token"] = accessToken
 	}
 }
 
@@ -84,6 +76,18 @@ func WithServiceVersion(version string) Option {
 	}
 }
 
+// WithHeaders configures OTLP/gRPC connection headers
+func WithHeaders(headers map[string]string) Option {
+	return func(c *Config) {
+		if c.Headers == nil {
+			c.Headers = make(map[string]string)
+		}
+		for k, v := range headers {
+			c.Headers[k] = v
+		}
+	}
+}
+
 // WithLogLevel configures the logging level for OpenTelemetry
 func WithLogLevel(loglevel string) Option {
 	return func(c *Config) {
@@ -110,7 +114,7 @@ func WithMetricExporterInsecure(insecure bool) Option {
 // WithResourceAttributes configures attributes on the resource
 func WithResourceAttributes(attributes map[string]string) Option {
 	return func(c *Config) {
-		c.resourceAttributes = attributes
+		c.ResourceAttributes = attributes
 	}
 }
 
@@ -184,18 +188,18 @@ const (
 )
 
 type Config struct {
-	SpanExporterEndpoint           string   `env:"OTEL_EXPORTER_OTLP_SPAN_ENDPOINT,default=ingest.lightstep.com:443"`
-	SpanExporterEndpointInsecure   bool     `env:"OTEL_EXPORTER_OTLP_SPAN_INSECURE,default=false"`
-	ServiceName                    string   `env:"LS_SERVICE_NAME"`
-	ServiceVersion                 string   `env:"LS_SERVICE_VERSION,default=unknown"`
-	MetricExporterEndpoint         string   `env:"OTEL_EXPORTER_OTLP_METRIC_ENDPOINT,default=ingest.lightstep.com:443"`
-	MetricExporterEndpointInsecure bool     `env:"OTEL_EXPORTER_OTLP_METRIC_INSECURE,default=false"`
-	MetricsEnabled                 bool     `env:"LS_METRICS_ENABLED,default=true"`
-	AccessToken                    string   `env:"LS_ACCESS_TOKEN"`
-	LogLevel                       string   `env:"OTEL_LOG_LEVEL,default=info"`
-	Propagators                    []string `env:"OTEL_PROPAGATORS,default=b3"`
-	MetricReportingPeriod          string   `env:"OTEL_EXPORTER_OTLP_METRIC_PERIOD,default=30s"`
-	resourceAttributes             map[string]string
+	SpanExporterEndpoint           string            `env:"OTEL_EXPORTER_OTLP_SPAN_ENDPOINT,default=ingest.lightstep.com:443"`
+	SpanExporterEndpointInsecure   bool              `env:"OTEL_EXPORTER_OTLP_SPAN_INSECURE,default=false"`
+	ServiceName                    string            `env:"LS_SERVICE_NAME"`
+	ServiceVersion                 string            `env:"LS_SERVICE_VERSION,default=unknown"`
+	Headers                        map[string]string `env:"OTEL_EXPORTER_OTLP_HEADERS"`
+	MetricExporterEndpoint         string            `env:"OTEL_EXPORTER_OTLP_METRIC_ENDPOINT,default=ingest.lightstep.com:443"`
+	MetricExporterEndpointInsecure bool              `env:"OTEL_EXPORTER_OTLP_METRIC_INSECURE,default=false"`
+	MetricsEnabled                 bool              `env:"LS_METRICS_ENABLED,default=true"`
+	LogLevel                       string            `env:"OTEL_LOG_LEVEL,default=info"`
+	Propagators                    []string          `env:"OTEL_PROPAGATORS,default=b3"`
+	MetricReportingPeriod          string            `env:"OTEL_EXPORTER_OTLP_METRIC_PERIOD,default=30s"`
+	ResourceAttributes             map[string]string
 	Resource                       *resource.Resource
 	logger                         Logger
 	errorHandler                   otel.ErrorHandler
@@ -212,11 +216,18 @@ func checkEndpointDefault(value, defValue string) error {
 	return nil
 }
 
+func accessToken(c Config) string {
+	if c.Headers == nil {
+		return ""
+	}
+	return c.Headers[lightstepAccessTokenHeader]
+}
+
 func validateConfiguration(c Config) error {
 	if len(c.ServiceName) == 0 {
 		serviceNameSet := false
 		for _, kv := range c.Resource.Attributes() {
-			if kv.Key == conventions.AttributeServiceName {
+			if kv.Key == semconv.ServiceNameKey {
 				if len(kv.Value.AsString()) > 0 {
 					serviceNameSet = true
 				}
@@ -228,7 +239,7 @@ func validateConfiguration(c Config) error {
 		}
 	}
 
-	accessTokenLen := len(c.AccessToken)
+	accessTokenLen := len(accessToken(c))
 	if accessTokenLen == 0 {
 		if err := checkEndpointDefault(c.SpanExporterEndpoint, DefaultSpanExporterEndpoint); err != nil {
 			return err
@@ -239,9 +250,11 @@ func validateConfiguration(c Config) error {
 		}
 	}
 
+	// TODO(@tobert) will probably break on some providers but seems fine for my use cases right now
 	if accessTokenLen > 0 && (c.AccessToken != "developer" && accessTokenLen != 32 && accessTokenLen != 84 && accessTokenLen != 104) {
 		return fmt.Errorf("invalid configuration: access token length incorrect. Ensure token is set correctly")
 	}
+
 	return nil
 }
 
@@ -269,60 +282,36 @@ type Launcher struct {
 	shutdownFuncs []func() error
 }
 
-// configurePropagators configures B3 propagation by default
-func configurePropagators(c *Config) error {
-	propagatorsMap := map[string]propagation.TextMapPropagator{
-		"b3":           b3.B3{InjectEncoding: b3.B3MultipleHeader},
-		"baggage":      propagation.Baggage{},
-		"tracecontext": propagation.TraceContext{},
-		"ottrace":      ot.OT{},
-	}
-	var props []propagation.TextMapPropagator
-	for _, key := range c.Propagators {
-		prop := propagatorsMap[key]
-		if prop != nil {
-			props = append(props, prop)
-		}
-	}
-	if len(props) == 0 {
-		return fmt.Errorf("invalid configuration: unsupported propagators. Supported options: b3,baggage,tracecontext,ottrace")
-	}
-	otel.SetTextMapPropagator(propagation.NewCompositeTextMapPropagator(
-		props...,
-	))
-	return nil
-}
-
 func newResource(c *Config) *resource.Resource {
 	r := resource.Environment()
 
 	hostnameSet := false
 	for iter := r.Iter(); iter.Next(); {
-		if iter.Attribute().Key == conventions.AttributeHostName && len(iter.Attribute().Value.Emit()) > 0 {
+		if iter.Attribute().Key == semconv.HostNameKey && len(iter.Attribute().Value.Emit()) > 0 {
 			hostnameSet = true
 		}
 	}
 
 	attributes := []attribute.KeyValue{
-		attribute.String(conventions.AttributeTelemetrySDKName, "launcher"),
-		attribute.String(conventions.AttributeTelemetrySDKLanguage, "go"),
-		attribute.String(conventions.AttributeTelemetrySDKVersion, version),
+		semconv.TelemetrySDKNameKey.String("launcher"),
+		semconv.TelemetrySDKLanguageGo,
+		semconv.TelemetrySDKVersionKey.String(version),
 	}
 
 	if len(c.ServiceName) > 0 {
-		attributes = append(attributes, attribute.String(conventions.AttributeServiceName, c.ServiceName))
+		attributes = append(attributes, semconv.ServiceNameKey.String(c.ServiceName))
 	}
 
 	if len(c.ServiceVersion) > 0 {
-		attributes = append(attributes, attribute.String(conventions.AttributeServiceVersion, c.ServiceVersion))
+		attributes = append(attributes, semconv.ServiceVersionKey.String(c.ServiceVersion))
 	}
 
-	for key, value := range c.resourceAttributes {
+	for key, value := range c.ResourceAttributes {
 		if len(value) > 0 {
-			if key == conventions.AttributeHostName {
+			if key == string(semconv.HostNameKey) {
 				hostnameSet = true
 			}
-			attributes = append(attributes, attribute.String(key, value))
+			attributes = append(attributes, semconv.HostNameKey.String(value))
 		}
 	}
 
@@ -331,7 +320,7 @@ func newResource(c *Config) *resource.Resource {
 		if err != nil {
 			c.logger.Debugf("unable to set host.name. Set OTEL_RESOURCE_ATTRIBUTES=\"host.name=<your_host_name>\" env var or configure WithResourceAttributes in code: %v", err)
 		} else {
-			attributes = append(attributes, attribute.String(conventions.AttributeHostName, hostname))
+			attributes = append(attributes, semconv.HostNameKey.String(hostname))
 		}
 	}
 
@@ -350,73 +339,18 @@ func newResource(c *Config) *resource.Resource {
 	return r
 }
 
-func newTraceExporter(accessToken, endpoint string, insecure bool) (*otlptrace.Exporter, error) {
-	headers := map[string]string{
-		"lightstep-access-token": accessToken,
-	}
-
-	secureOption := otlptracegrpc.WithTLSCredentials(credentials.NewClientTLSFromCert(nil, ""))
-	if insecure {
-		secureOption = otlptracegrpc.WithInsecure()
-	}
-	return otlptrace.New(
-		context.Background(),
-		otlptracegrpc.NewClient(
-			secureOption,
-			otlptracegrpc.WithEndpoint(endpoint),
-			otlptracegrpc.WithHeaders(headers),
-			otlptracegrpc.WithCompressor(gzip.Name),
-		),
-	)
-}
-
-func newMetricsExporter(accessToken, endpoint string, insecure bool) (*otlpmetric.Exporter, error) {
-	headers := map[string]string{
-		"lightstep-access-token": accessToken,
-	}
-
-	secureOption := otlpmetricgrpc.WithTLSCredentials(credentials.NewClientTLSFromCert(nil, ""))
-	if insecure {
-		secureOption = otlpmetricgrpc.WithInsecure()
-	}
-	return otlpmetric.New(
-		context.Background(),
-		otlpmetricgrpc.NewClient(
-			secureOption,
-			otlpmetricgrpc.WithEndpoint(endpoint),
-			otlpmetricgrpc.WithHeaders(headers),
-			otlpmetricgrpc.WithCompressor(gzip.Name),
-		),
-	)
-}
-
 func setupTracing(c Config) (func() error, error) {
 	if c.SpanExporterEndpoint == "" {
 		c.logger.Debugf("tracing is disabled by configuration: no endpoint set")
 		return nil, nil
 	}
-	spanExporter, err := newTraceExporter(c.AccessToken, c.SpanExporterEndpoint, c.SpanExporterEndpointInsecure)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create span exporter: %v", err)
-	}
-
-	bsp := trace.NewBatchSpanProcessor(spanExporter)
-	tp := trace.NewTracerProvider(
-		trace.WithSampler(trace.AlwaysSample()),
-		trace.WithSpanProcessor(bsp),
-		trace.WithResource(c.Resource),
-	)
-
-	if err = configurePropagators(&c); err != nil {
-		return nil, err
-	}
-
-	otel.SetTracerProvider(tp)
-
-	return func() error {
-		_ = bsp.Shutdown(context.Background())
-		return spanExporter.Shutdown(context.Background())
-	}, nil
+	return pipelines.NewTracePipeline(pipelines.PipelineConfig{
+		Endpoint:    c.SpanExporterEndpoint,
+		Insecure:    c.SpanExporterEndpointInsecure,
+		Headers:     c.Headers,
+		Resource:    c.Resource,
+		Propagators: c.Propagators,
+	})
 }
 
 type setupFunc func(Config) (func() error, error)
@@ -426,51 +360,13 @@ func setupMetrics(c Config) (func() error, error) {
 		c.logger.Debugf("metrics are disabled by configuration: no endpoint set")
 		return nil, nil
 	}
-	metricExporter, err := newMetricsExporter(c.AccessToken, c.MetricExporterEndpoint, c.MetricExporterEndpointInsecure)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create metric exporter: %v", err)
-	}
-
-	period := controller.DefaultPeriod
-	if c.MetricReportingPeriod != "" {
-		period, err = time.ParseDuration(c.MetricReportingPeriod)
-		if err != nil {
-			return nil, fmt.Errorf("invalid metric reporting period: %v", err)
-		}
-		if period <= 0 {
-
-			return nil, fmt.Errorf("invalid metric reporting period: %v", c.MetricReportingPeriod)
-		}
-	}
-	pusher := controller.New(
-		processor.New(
-			selector.NewWithInexpensiveDistribution(),
-			metricExporter,
-		),
-		controller.WithExporter(metricExporter),
-		controller.WithResource(c.Resource),
-		controller.WithCollectPeriod(period),
-	)
-
-	if err = pusher.Start(context.Background()); err != nil {
-		return nil, fmt.Errorf("failed to start controller: %v", err)
-	}
-
-	provider := pusher.MeterProvider()
-
-	if err = runtimeMetrics.Start(runtimeMetrics.WithMeterProvider(provider)); err != nil {
-		return nil, fmt.Errorf("failed to start runtime metrics: %v", err)
-	}
-
-	if err = hostMetrics.Start(hostMetrics.WithMeterProvider(provider)); err != nil {
-		return nil, fmt.Errorf("failed to start host metrics: %v", err)
-	}
-
-	metricglobal.SetMeterProvider(provider)
-	return func() error {
-		_ = pusher.Stop(context.Background())
-		return metricExporter.Shutdown(context.Background())
-	}, nil
+	return pipelines.NewMetricsPipeline(pipelines.PipelineConfig{
+		Endpoint:        c.MetricExporterEndpoint,
+		Insecure:        c.MetricExporterEndpointInsecure,
+		Headers:         c.Headers,
+		Resource:        c.Resource,
+		ReportingPeriod: c.MetricReportingPeriod,
+	})
 }
 
 func ConfigureOpentelemetry(opts ...Option) Launcher {
@@ -481,6 +377,15 @@ func ConfigureOpentelemetry(opts ...Option) Launcher {
 		c.logger.Debugf("configuration")
 		s, _ := json.MarshalIndent(c, "", "\t")
 		c.logger.Debugf(string(s))
+	}
+
+	if c.Headers == nil {
+		c.Headers = map[string]string{}
+	}
+
+	token := os.Getenv("LS_ACCESS_TOKEN")
+	if len(token) > 0 && len(c.Headers[lightstepAccessTokenHeader]) == 0 {
+		c.Headers[lightstepAccessTokenHeader] = token
 	}
 
 	err := validateConfiguration(c)
@@ -495,6 +400,7 @@ func ConfigureOpentelemetry(opts ...Option) Launcher {
 	ls := Launcher{
 		config: c,
 	}
+
 	for _, setup := range []setupFunc{setupTracing, setupMetrics} {
 		shutdown, err := setup(c)
 		if err != nil {