+++ CHANGELOG.md
@@ -8,11 +8,21 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
-## [1.0.0-RC1](https://github.com/lightstep/otel-launcher-go/releases/tag/v1.0.0-RC1) - 2021-06-21
+## 💡 Enhancements 💡
+
+- use semconv package instead of collector package (#69)
+- update OpenTelemetry SDK dependencies to 1.4.0 (#112)
+- update OpenTelemetry metrics to 0.27.0 (#114)
+
+## [1.0.0](https://github.com/lightstep/otel-launcher-go/releases/tag/v1.0.0) - 2021-09-22
 
 ### Changed
-- Update OpenTelemetry tracing dependencies to 1.0.0-RC1
-- Update OpenTelemetry metrics dependencies to 0.21.0
+- Update OpenTelemetry tracing dependencies to 1.0.0
+## [1.0.0-RC3](https://github.com/lightstep/otel-launcher-go/releases/tag/v1.0.0-RC3) - 2021-09-15
+
+### Changed
+- Update OpenTelemetry tracing dependencies to 1.0.0-RC3
+- Update OpenTelemetry metrics dependencies to 0.23.0
 - Remove support for deprecated `OTEL_RESOURCE_LABELS` environment variable, as
   it has been replaced by `OTEL_RESOURCE_ATTRIBUTES` upstream.